# net: Improve performance of isIPv4 and isIPv6