Corresponding PR for https://github.com/nodejs/node/pull/49568

```js
'use strict'

const { isIPv4: isIPv4builtIn } = require('net')
const benchmark = require('benchmark')
const suite = new benchmark.Suite()

const ipMax = '255.255.255.255'
const ipMin = '0.0.0.0'
const invalid = '0.0.0.0.0'

const v4Seg = '(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9][0-9]|[0-9])';
const v4Str = `(?:${v4Seg}\\.){3}${v4Seg}`;
const IPv4Reg = new RegExp(`^${v4Str}$`);

const isIPv4R = IPv4Reg.test.bind(IPv4Reg)

suite.add('RE (bound)', function () {
  isIPv4R(ipMin)
  isIPv4R(ipMax)
  isIPv4R(invalid)
})

suite.add('RE wrapped', function () {
  IPv4Reg.test(ipMin)
  IPv4Reg.test(ipMax)
  IPv4Reg.test(invalid)
})

suite.add('builtIn', function () {
  isIPv4builtIn(ipMin)
  isIPv4builtIn(ipMax)
  isIPv4builtIn(invalid)
})

suite.on('cycle', function onCycle(event) {
  console.log(String(event.target))
})

suite.run({ async: false })
```

```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/proxy-addr$ bun run ./benchmark/is-ipv4.js 
RE (bound) x 194,995,252 ops/sec ±56.36% (25 runs sampled)
RE wrapped x 218,862,836 ops/sec ±54.86% (29 runs sampled)
builtIn x 26,586,761 ops/sec ±4.99% (26 runs sampled)
```

It has about 55% deviation. If I change minSamples to 100 then I get about 

```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/proxy-addr$ bun run ./benchmark/is-ipv4.js 
RE (bound) x 73,035,238 ops/sec ±8.13% (107 runs sampled)
RE wrapped x 74,912,245 ops/sec ±8.86% (107 runs sampled)
builtIn x 18,630,794 ops/sec ±2.12% (116 runs sampled)
```

So maybe some benchmarking from your side would help to verify the benchmark gains?